Share the bounds the three engines would otherwise each copy
The list bound (notify.Listed) and the notification job shape (notify.SnapshotJob)
were about to exist three times. They are rules, not utilities: a private copy per
engine is a copy that drifts, and one page reads all three.

The batching loop itself stays duplicated between reminder and nudge — a shared
engine would have to be generic over the delivery row, the ledger and the message
type, which costs more than the ~80 lines it saves. The seam is noted where the
next reader will look.
