+++ internal/engage/reminder/engine.go
@@ -2,6 +2,7 @@ package reminder
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"log"
@@ -373,6 +374,16 @@ func (r *Runner) recordNotification(ctx context.Context, b *batch) {
 	if _, err := r.store.RecordNotification(ctx, arg); err != nil {
 		log.Printf("reminder: record notification for user %d: %v", b.info.UserID, err)
 	}
+}
+
+// jobsSnapshot is the job list a multi-job batch records for its notification's own
+// page, in the shape notify owns because one page renders every engine's rows.
+func jobsSnapshot(ms []ReminderMessage) json.RawMessage {
+	jobs := make([]notify.SnapshotJob, len(ms))
+	for i, m := range ms {
+		jobs[i] = notify.SnapshotJob{Title: m.JobTitle, Company: m.Company, Slug: m.Slug}
+	}
+	return notify.JobsSnapshot(jobs)
 }
 
 // release drops the lease on a reminder so it is retried promptly on a later pass.
